requirements1.txt
altair==5.5.0
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
captcha==0.6.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
et_xmlfile==2.0.0
extra-streamlit-components
filetype==1.2.0
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python
narwhals==1.23.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27 
pytz==2024.2
# pywin32==310
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.37
streamlit
streamlit-aggrid==1.1.0
streamlit-authenticator
streamlit-calendar
streamlit-cookies-controller
streamlit-cookies-manager
streamlit-js-eval
streamlit-lottie
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
streamlit_extras
ffmpeg-python
pymysql
